MAJOR refactoring. Made the code pattern more consistent, modular, and clear. Hopefully further code changes will be much easier/quicker.
Also started experimenting with sprite resources. The game now loads a single sprite sheet and uses clip ("crop" in canvas/Kineticjs) to get specific card images. This seems to be working okay in Chrome and IE. However, Firefox seems to be having some big problems with it. Framerate on Firefox has dropped to unbearable amounts. Will definitely have to do some more research/experimenting with this.